HDDS-10071. Merge TestContainerStateMachineIdempotency into TestContainerOperations (#6813)
