+++ hadoop-ozone/integration-test/dev-support/findbugsExcludeFile.xml
@@ -41,7 +41,7 @@
     <Bug pattern="DLS_DEAD_LOCAL_STORE" />
   </Match>
   <Match>
-    <Class name="org.apache.hadoop.ozone.TestContainerStateMachineIdempotency"/>
+    <Class name="org.apache.hadoop.ozone.TestContainerOperations"/>
     <Bug pattern="RV_RETURN_VALUE_IGNORED_NO_SIDE_EFFECT" />
   </Match>
   <Match>
